refactor: split entities/switch.py into themed modules
