+++ custom_components/glinet_router/entities/switch_core.py
@@ -1,0 +1,80 @@
+from __future__ import annotations
+
+import logging
+from typing import Any
+
+from ..hub import GLinetHub
+from ..models import WifiInterface
+from .switch_base import GLinetSwitchBase
+
+_LOGGER = logging.getLogger(__name__)
+
+
+class WifiApSwitch(GLinetSwitchBase):
+    def __init__(self, hub: GLinetHub, iface_name: str, iface: WifiInterface) -> None:
+        super().__init__(hub)
+        self._iface_name = iface_name
+        self._iface = iface
+
+    @property
+    def unique_id(self) -> str:
+        return f"glinet_switch/{self._hub.device_mac}/iface_{self._iface_name}"
+
+    @property
+    def name(self) -> str:
+        return self._iface.ssid or self._iface.name
+
+    @property
+    def is_on(self) -> bool | None:
+        self._iface = self._hub.wifi_interfaces.get(self._iface_name, self._iface)
+        return self._iface.enabled
+
+    @property
+    def extra_state_attributes(self) -> dict[str, str | bool]:
+        return {
+            "interface": self._iface.name,
+            "guest": self._iface.guest,
+            "ssid": self._iface.ssid,
+            "hidden": self._iface.hidden,
+            "encryption": self._iface.encryption,
+        }
+
+    async def async_turn_on(self, **_: Any) -> None:
+        try:
+            await self._hub.set_wifi_interface_enabled(self._iface_name, True)
+        except OSError:
+            _LOGGER.exception("Unable to enable WiFi interface %s", self._iface_name)
+            return
+        await self._hub.async_request_refresh()
+
+    async def async_turn_off(self, **_: Any) -> None:
+        try:
+            await self._hub.set_wifi_interface_enabled(self._iface_name, False)
+        except OSError:
+            _LOGGER.exception("Unable to disable WiFi interface %s", self._iface_name)
+            return
+        await self._hub.async_request_refresh()
+
+
+class LedSwitch(GLinetSwitchBase):
+    _attr_icon = "mdi:led-on"
+
+    @property
+    def unique_id(self) -> str:
+        return f"glinet_switch/{self._hub.device_mac}/led"
+
+    @property
+    def name(self) -> str:
+        return "System LED"
+
+    @property
+    def is_on(self) -> bool | None:
+        return self._hub.led_enabled
+
+    async def async_turn_on(self, **_: Any) -> None:
+        await self._hub.set_led_enabled(True)
+        await self._hub.async_request_refresh()
+
+    async def async_turn_off(self, **_: Any) -> None:
+        await self._hub.set_led_enabled(False)
+        await self._hub.async_request_refresh()
